Match hermite/quake/tv_mode/cString/xml/room_tex; partial rnd/fade/file/texture/geometry; engine headers (fade, quake, texture, mes, gx FIFO); more idioms

+++ include/filter.h
@@ -48,7 +48,7 @@ void Filter0bTrans();
 
 extern "C" {
 // filter00.cpp: additive radial blur request (highest priority wins)
-void Filter00SetAddSpread(u8 pri, int on, u8 r, u8 g, u8 b, u8 rate, u8 type, int num, f32 cx, f32 cy, f32 pow);
+void Filter00SetAddSpread(u32 pri, int on, u8 r, u8 g, u8 b, u8 rate, u8 type, u32 num, f32 cx, f32 cy, f32 pow);
 // filter03.cpp
 void Filter03SetParam(int level, u8 r, u8 g, u8 b, u8 pri, int flag);
 }
